refactor to model core crate
