+++ crates/thyllore-model-core/src/lib.rs
@@ -1,0 +1,5 @@
+mod skeleton;
+mod skin;
+
+pub use skeleton::*;
+pub use skin::*;
